Move Screen to commonMain
